fix(tools): share terminal output cap
