+++ src/hosts/uiHosts.ts
@@ -1,6 +1,3 @@
-// Third-party imports
-import stripAnsi from 'strip-ansi';
-
 export interface DiffSource {
   filePath: string;
 }
@@ -92,25 +89,6 @@ export interface TerminalRunRequest {
   command: string;
   /** Hard cap on how long to wait for captured execution. */
   timeoutMs: number;
-}
-
-/**
- * Length cap (chars) for {@link TerminalRunResult.output}. Apply it through
- * {@link truncateTerminalOutput} so agents see identical truncation regardless
- * of where a setup command ran.
- */
-export const TERMINAL_OUTPUT_MAX_CHARS = 12_000;
-
-/**
- * The shared ANSI-stripped, length-capped tail every host's `TerminalRunner`
- * returns in {@link TerminalRunResult.output}. Strips control sequences, then
- * keeps the trailing {@link TERMINAL_OUTPUT_MAX_CHARS} characters.
- */
-export function truncateTerminalOutput(output: string): string {
-  const stripped = stripAnsi(output);
-  return stripped.length > TERMINAL_OUTPUT_MAX_CHARS
-    ? stripped.slice(-TERMINAL_OUTPUT_MAX_CHARS)
-    : stripped;
 }
 
 export interface TerminalRunResult {
