+++ src/test-kernel/tools/setup/SendToTerminalTool.vitest.ts
@@ -8,7 +8,6 @@ import { strict as assert } from 'node:assert';
 import { describe, it, beforeAll } from 'vitest';
 
 // Local imports
-import { TERMINAL_OUTPUT_MAX_CHARS } from '@hosts/uiHosts';
 import type { ConfigProvider } from '@platform/interfaces';
 import { BASH_APPROVAL_CONFIG_KEY } from '@shared/schemas';
 import { installPlatform } from '@test/support/setupPlatform';
@@ -17,6 +16,7 @@ import {
   setSetupPlatform,
   type TerminalRunResult,
 } from '@tools/setup/platform';
+import { TERMINAL_OUTPUT_MAX_CHARS } from '@utils/core/terminalOutput';
 
 // Local file imports
 import { createFakeSetupPlatform } from './fixtures';
